builder/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
docker
requests
google-api-core
google-cloud-build
google-cloud-container
google-cloud-storage
google-cloud-logging
google-cloud-artifact-registry==1.9.0
pymupdf
numpy
